Created facade layer fo Objekt
